:books: Major update to documentation
* :books: The documentation for the classification functions has been rewritten.

All descriptions is on the form "This foo(), also known as, returns ..." with wikipedia links. The details sections have been rewritten for all
functions to include a discussion about correct specification of factors.

The mathmatics behind the funtions have been moved to a custom section called "calculation". For the next commit, all mathmatical formulas will be updated and checked.