CRAN Preperation :fire:  (#35)
## :books: What?

This PR is the first step towards a CRAN submission. The actual
submission will happen around March 2025, or when the package is deemed
ready for the wider public. The PR introduces the following essential
changes.

1. The package is now being checked locally with `--as-cran` - it was
originally planned to implement it in github actions, but this is not
possible due to latex dependencies
2. Added programmatic rendering of README and NEWS via `make document` -
this might be a better idea to implement via pre-commit hooks, however.

> [!WARNING]
>
> It seems that the `prROC.plot()` and `ROC.plot()` are broken. See NEWS
files.