Updated Workflow - See commit message :books: (#41)
## :books: What?
* Performance Tests: All performance tests have been appropriately split
between the type of test. This enables tests to be run individually in
an easier fashion. The dependency on the setup.R-script in the
test-folder has been removed. All tests can now be run independently.
* MAKE-file: The make-file have been rearranged as follows:

1. A new commmand "performance" will run all performance scripts and
overwrite the existing file.
2. The "build" commmand now builds the {pkgdown} website and renders the
NEWS and README after each run.
3. The "check" command now installs the package after checks too. This
is mainly done for debugging purposes if tests fails, to avoid having to
rebuild everything.

* Tools: The render functions have been moved to its own script as it
make sense to build README and NEWS AFTER building the package. Not
before.