New utility function and minor changes :hammer:  (#48)
## :books: What?

* **New covariance matrix function:** This function is used in the
`ccc()`-function. The function is written with OpenMP and is the first
step towards full OpenMP support..
* **Updated Unit-tests:** The unit tests are now accounts for floating
point differences via relative and absolute tolerances. The size of the
data used for the tests have been reduces to avoid masking bugs that is
basically invisible in large sample sizes.
* **Skipping tests:** The `ccc()` for weighted with corrections are
being skipped for now. This is the third attempt to fix the error - its
different from scikit-learn, yardsttick and manual calculations. The
reason for this is unkown.
* **Ohter:** Other minor fixes to the package structure (See commit
history)