:books: Function references (#50)
## :books: What?

This PR adds a new set of documentation tools and a new part to the
online documentation: "Functions". The new tools builds the .qmd from
the .Rd files. This functionality builds on
[{altdocs}](https://github.com/etiennebacher/altdoc) with my own
modification so it suits the overall structure of the docs.

See PR https://github.com/serkor1/SLmetrics/pull/50 for more details.