:fist_right: :fist_left:  Enable collaboration (#54)
## :books: What?

The migration to .qmd files comes with many benefits, and some quirks.
When forking {SLmetrics} it is not possible to `$make build-meta`
without `--cache-refresh`. To alleviate this the following changes have
been implemented:

* **Add all artifacts to repository:** The repository now includes all
related artifacts created by quarto related to README and NEWS.
* **Updated build-meta command:** The command can now be executed with
`RESET=true` to refresh the cache. It calls `quarto --cache-refresh`
under the hood.

> [!NOTE]
>
> It introduces 105 changed files. This might clutter the repository
history, so it should be considered whether it should be rendered by
Github actions instead - or at least remember to commit it separately.