:hammer: Refactored OpenMP availability check
The original implementation of checking OpenMP availability were fundamentally flawed. #ifdef _OpenMP will be TRUE if there is a OpenMP **ready** compiler which is the case for clang and LLVM. To check if its correctly *setup* it should return more than a single thread on most modern systems (if not all); so this is one way to test it. If there is a another and more robust approach to this it will propably come to me at some point.