:hammer:  Refactored confusion matrix workflow
* The classification::confusion_matrix is now **only** dedicated to Arma::Mat and is used for internal calculations.
    - The class includes methods to calculate FP, TP, FN and TN to calculate various metrics.
    - The class is missing micro and na_rm arguments.
* The derived class metric::confusion_matrix is for exporting confusion matrices to the R-side.

NOTE: WIth this setup Lambda-functions can be made with appropriate classes to reduce the number of code. A similar method was done for regression tasks, but it introduced a significant overhead and later removed. The confusion matrix at the moment is 3x faster than before, so the additional cost of lambda-function may be considered offset. Time will tell.