:hammer: Updated Likelihood Ratio functions
* The likelihood ratio functions (dor, nlr, plr) have been rewritten so
they only return a double. This is consistent behaviour with everything
else and mathematically correct.

* The unit-test setup.R have been updated such that the imbalanced
classification is only relevant for k > 2. Before this change it was
setting one of the classes to probability 0 which essentially means that
a classifier only predicts a single class. Id assume that people check
such things before throwing them in the grinder.