:rocket: Micro-optimization of regression headers (#110)
## :books: What?

This PR is focused on streamlining the regression headers, and apply micro-optimizations. This PR focuses on the following changes:

* **noexcept**: the *noexcept* keyword *may* grant the compiler the freedom to aggressively optimize during compiling see, for example, [here](https://stackoverflow.com/questions/10787766/when-should-i-really-use-noexcept).
* **__restrict__:** Same story as above. It should enable aggressive optimizations during compilation - but local tests does not confirm this. But again, what do I know. The results may be a product of `R`, my compiler, my computer or the appended flags.
* **Logic:** Where applicable the logic has been improved and/or streamlined, so it is easy to navigate the source code for debugging. 

### :warning: Disclaimer :warning: 

The new setup effectively puts an end to OpenMP support unless the source code is refactored (again, again), which is *not* gonna happen. C++23 has a really cool zip-feature that can be used with `std::ranges::transform` which can has built-in parallel policies - but the current state of the configre scripts does not allow for C++23. This is for later implementations, if there is demand for it - which is highly unlikely.