:hammer_and_wrench: Refactored Entropy functions (#111)
## :books: What?

This `PR` is focused on the entropy-functions, and migrates all functions to the same structure as the classification and regression functions; a base-class and a derived class. The solution is not as elegant as I would want it to be but, as with everything else, it is a learning process. This PR changes the following:

* **Reworked parameters:** The argument `base` has been removed, and replaced with `normalized` which normalizes the calculated entropy across chosen dimensions.
* **Poisson Logloss:** The poisson logloss now accepts a vector of probabilities instead of a column matrix.
* **Logloss:** The logloss function now checks for extreme values towards 0 similar to the poisson logloss.
* **Updated documentation:** All entropy based functions have had added proper examples.